Moved RosterManagerViewComponent to /Types
